requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
google-generativeai==0.8.2
duckduckgo-search>=6.3.0
slowapi==0.1.9
pydantic>=2.9.2
pydantic-settings>=2.6.0
python-dotenv==1.0.1
